Drop dependency on nix
Rewrite the vfio backend to rely only on libc, and swap the dependency
on nix (licensed as "MIT") by a dependency on libc (licensed as "MIT OR
Apache-2.0").

This gets us to a point where both our own code and all of our (non-dev)
dependencies are licensed as "MIT OR Apache-2.0".

+++ src/backends/vfio/containers.rs
@@ -242,7 +242,7 @@ impl VfioContainer {
 
         // check extension
 
-        if unsafe { vfio_check_extension(fd, VFIO_TYPE1v2_IOMMU as i32)? } != 1 {
+        if unsafe { vfio_check_extension(fd, VFIO_TYPE1v2_IOMMU as usize)? } != 1 {
             return Err(io::Error::new(ErrorKind::InvalidInput, "TODO"));
         }
 
@@ -254,7 +254,7 @@ impl VfioContainer {
 
         // enable IOMMU
 
-        unsafe { vfio_set_iommu(fd, VFIO_TYPE1v2_IOMMU as i32)? };
+        unsafe { vfio_set_iommu(fd, VFIO_TYPE1v2_IOMMU as usize)? };
 
         // get IOMMU info
